requirements.txt
pandas
SQLAlchemy
psycopg2-binary
python-dotenv
requests
psycopg2-binary
python-dateutil
numpy